[Tests] clean and small refactor.:
