+++ test/verifier/verifyPresentationHelper.test.ts
@@ -76,102 +76,6 @@ const copyCredentialObj = (credential: JSONObj, elemName: string, elemValue = ''
   return (newCred);
 };
 
-const populateMockData = (): JSONObj => {
-  const context: string[] = ['https://www.w3.org/2018/credentials/v1'];
-  const type: string[] = ['VerifiablePresentation'];
-
-  const credentialIssuerDid = 'did:unum:7fc1753e-cdb7-428a-b6ce-eefc0e3634e5'; // could be dummyIssuerDid?
-  const verifiableCredentialObj =
-      {
-        '@context': [
-          'https://www.w3.org/2018/credentials/v1'
-        ],
-        credentialStatus: {
-          id: 'https://api.dev-unumid.org//credentialStatus/f8287c1e-0c56-460a-92af-5519f5c10cbf',
-          type: 'CredentialStatus'
-        },
-        credentialSubject: {
-          id: 'did:unum:5f5eb3dd-d0e0-4356-bfdd-96bc1393c705',
-          test: 'test'
-        },
-        issuer: credentialIssuerDid,
-        type: [
-          'VerifiableCredential',
-          'UsernameCredential'
-        ],
-        id: 'f8287c1e-0c56-460a-92af-5519f5c10cbf',
-        issuanceDate: '2021-01-09T02:23:54.844Z',
-        expirationDate: '2022-01-09T00:00:00.000Z',
-        proof: {
-          created: '2021-01-09T02:23:54.844Z',
-          type: 'secp256r1Signature2020',
-          verificationMethod: 'did:unum:7fc1753e-cdb7-428a-b6ce-eefc0e3634e5',
-          proofPurpose: 'AssertionMethod',
-          signatureValue: '381yXZCEPSC9NB2smArjiBtvnGL6LZ2yAUW1qLQfhuZSyeQiCyrFRqkxfPoa1gaLaScR7cFVJmguo1v1JKYH6uEU4Zd32D9C',
-          unsignedValue: '{"@context":["https://www.w3.org/2018/credentials/v1"],"credentialStatus":{"id":"https://api.dev-unumid.org//credentialStatus/f8287c1e-0c56-460a-92af-5519f5c10cbf","type":"CredentialStatus"},"credentialSubject":{"id":"did:unum:5f5eb3dd-d0e0-4356-bfdd-96bc1393c705","username":"Analyst-Shoes-278"},"expirationDate":"2022-01-09T00:00:00.000Z","id":"f8287c1e-0c56-460a-92af-5519f5c10cbf","issuanceDate":"2021-01-09T02:23:54.844Z","issuer":"did:unum:7fc1753e-cdb7-428a-b6ce-eefc0e3634e5","proof":{"created":"2021-01-09T02:23:54.844Z","proofPurpose":"AssertionMethod","type":"secp256r1Signature2020","verificationMethod":"did:unum:7fc1753e-cdb7-428a-b6ce-eefc0e3634e5"},"type":["VerifiableCredential","UsernameCredential"]}'
-        }
-      };
-  const verifiableCredential = [verifiableCredentialObj];
-  const verifiableCredentialString = [stringify(verifiableCredentialObj)];
-
-  const presentationRequestUuid = '0cebee3b-3295-4ef6-a4d6-7dfea413b3aa';
-  const invalidProof: JSONObj = {
-    created: '2020-09-03T18:50:52.105Z',
-    signatureValue: 'iTx1CJLYue7vopUo2fqGps3TWmxqRxoBDTupumLkaNp2W3UeAjwLUf5WxLRCRkDzEFeKCgT7JdF5fqbpvqnBZoHyYzWYbmW4YQ',
-    unsignedValue: stringify({}),
-    type: 'secp256r1Signature2020',
-    verificationMethod: 'did:unum:3ff2f020-50b0-4f4c-a267-a9f104aedcd8',
-    proofPurpose: 'AssertionMethod'
-  };
-  const authHeader = 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoidmVyaWZpZXIiLCJ1dWlkIjoiM2VjYzVlZDMtZjdhMC00OTU4LWJjOTgtYjc5NTQxMThmODUyIiwiZGlkIjoiZGlkOnVudW06ZWVhYmU0NGItNjcxMi00NTRkLWIzMWItNTM0NTg4NTlmMTFmIiwiZXhwIjoxNTk1NDcxNTc0LjQyMiwiaWF0IjoxNTk1NTI5NTExfQ.4iJn_a8fHnVsmegdR5uIsdCjXmyZ505x1nA8NVvTEBg';
-  const verifier = 'did:unum:dd407b1a-ee7f-46a2-af2a-ccbb48cbb0dc';
-
-  const subjectDid = 'did:unum:3ff2f020-50b0-4f4c-a267-a9f104aedcd8';
-  const unsignedPresentation = {
-    '@context': ['https://www.w3.org/2018/credentials/v1'],
-    presentationRequestUuid,
-    verifiableCredential: [verifiableCredential],
-    type: ['VerifiablePresentation']
-  };
-
-  const signatureValue = sign(unsignedPresentation, dummyEccPrivateKey, 'base58');
-
-  const proof: Proof = {
-    created: new Date().toISOString(),
-    signatureValue,
-    unsignedValue: stringify(unsignedPresentation),
-    type: 'secp256r1Signature2020',
-    verificationMethod: subjectDid,
-    proofPurpose: 'assertionMethod'
-  };
-
-  const presentation = {
-    ...unsignedPresentation,
-    proof
-  };
-
-  const credRequest: CredentialRequest = {
-    type: 'UsernameCredential',
-    issuers: [credentialIssuerDid],
-    required: true
-  };
-  const credentialRequests = [credRequest];
-
-  return ({
-    context,
-    type,
-    verifiableCredential,
-    verifiableCredentialString,
-    presentationRequestUuid,
-    proof,
-    invalidProof,
-    authHeader,
-    verifier,
-    presentation,
-    credentialRequests
-  });
-};
-
 describe('verifyPresentationHelper - Success Scenario with verifiableCredentialsString', () => {
   let response: UnumDto<VerifiedStatus>;
   let verStatus: boolean;
